CSharp Algorithm Comparison
