--- Movegen_dotNET/Sliders/ISlider.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Movegen_dotNET.Sliders
-{
-    public interface ISlider
-    {
-
-    }
-}
