+++ QBB.hpp
